Move the Geode-Projects form into its plugin
