+++ coldfront/plugins/geode_project/views.py
@@ -1,3 +1,13 @@
-from django.shortcuts import render
+from coldfront.plugins.geode_project.forms import GeodeProjectForm
 
-# Create your views here.
+
+class GeodeProjectView:
+    form_class = GeodeProjectForm
+    template_name = 'geode_project/geodeproject.html'
+
+    def form_valid(self, form):
+        http_response = super().form_valid(form)
+        self.allocation_obj.end_date = None
+        self.allocation_obj.save()
+
+        return http_response
